Change the behavior of array accessors for graphs
$g[ 'A' ] = $attrs;          => add node "A" with attributes from $attrs
$g[ ['A', 'B] ] = $attrs;    => add nodes "A" and "B" with attributes
                                from $attrs
$g[ ['A' => 'B'] ] = $attrs; => add an edge from "A" to "B" using attributes
                                from $attrs

$g[] = 'A';                  => add node "A" without any attribute
$g[] = ['A', 'B'];           => add nodes "A" and "B" without any attribute
$g[] = ['A' => 'B'];         => add an edge from "A" to "B" without any
                                attribute

Some of these can be mixed together, e.g.
$g[ ['A' => 'B', 'C'] ] = $attrs;
$g[] = ['A' => 'B', 'C'];

However, keep in mind that duplicate keys are not allowed inside PHP's
associative arrays, which may lead to unexpected results.

For example, the following line will add a single edge from "A" to "C"
because the key for the second entry ('A') collides with that of the
first entry:
  $g[] = ['A' => 'B', 'A' => 'C'];

This is probably not what the developper intended.